[S08P12A804-37] Feat/BE: Add logout Api & jwt token

+++ backend/src/main/java/com/ssafy/kkini/repository/RefreshTokenRepository.java
@@ -6,8 +6,8 @@ import org.springframework.stereotype.Repository;
 
 @Repository
 public interface RefreshTokenRepository extends JpaRepository<RefreshToken, Long> {
-    RefreshToken findByUserId(int userId);
+    RefreshToken findByUserId(Long userId);
 
     RefreshToken findByRefreshToken(String refreshToken);
-    RefreshToken findByUserIdAndRefreshToken(int userId, String refreshToken);
+    RefreshToken findByUserIdAndRefreshToken(Long userId, String refreshToken);
 }
