+++ backend/src/main/java/com/ssafy/kkini/repository/UserRepository.java
@@ -1,6 +1,8 @@
 package com.ssafy.kkini.repository;
 import com.ssafy.kkini.entity.User;
 import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.data.jpa.repository.Query;
+import org.springframework.data.repository.query.Param;
 import org.springframework.stereotype.Repository;
 
 import java.util.Optional;
@@ -14,5 +16,9 @@ public interface UserRepository extends JpaRepository<User, Long> {
 
     Optional<User> findByEmailAndProviderId(String email, String providerId);
 
-    Optional<User> findAllById(int userId);
+    Optional<User> findAllById(Long userId);
+
+//    @Query(value = "select u from User u where u.email = :email and u.password = :password")
+//    User login(@Param("email") String email, @Param("password") String password);
+
 }
